redesign home: first draft, resolves #21
