+++ src/assets/scss/layouts/_index.scss
@@ -3,7 +3,7 @@
   @extend %padding-regular;
   background-color: $header-desc-background-color;
   background-size: cover;
-  color: $header-desc-text-color;
+  color: $brand-black;
   display: inline-block;
   font-size: 1.2rem;
   text-align: center;
