+++ src/components/CourseList.jsx
@@ -2,16 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import urlJoin from 'url-join';
 
-const dateOptions = {
-  year: 'numeric',
-  month: 'long',
-  day: 'numeric',
-};
+import CourseItem from './CourseItem';
 
-const timeOptions = {
-  hour: '2-digit',
-  minute: '2-digit',
-};
+const courseRenderer = (course, index) => (
+  <CourseItem course={course} key={index} />
+);
 
 export default class CourseList extends Component {
   static propTypes = {
@@ -22,56 +17,10 @@ export default class CourseList extends Component {
     courses: [],
   };
 
-  constructor(props) {
-    super(props);
-    this.courseRenderer = this.courseRenderer.bind(this);
-    this.authorsRenderer = this.authorsRenderer.bind(this);
-  }
-
-  authorsRenderer(author) {
-    return (
-      <a
-        className="author-link"
-        href={urlJoin('http://blog.eleven-labs.com/authors/', author.username)}
-        key={author.username}
-      >
-        {author.name}
-      </a>
-    );
-  }
-
-  courseRenderer(course, index) {
-    const date = new Date(course.date);
-
-    return (
-      <div className="posts-teaser slice" key={index}>
-        <div className="container">
-          <h2 className="posts-title">
-            <a className="no-link-style" href={urlJoin('/course/', course.permalink)}>
-              {course.title}
-            </a>
-          </h2>
-
-          <time className="posts-date meta">
-            <span className="meta-content">
-              {date.toLocaleString('fr-FR', dateOptions)}
-              {' par '}
-              {course.authors.map(this.authorsRenderer)}
-            </span>
-          </time>
-
-          <p>{course.description}</p>
-
-          <a className="button" href={urlJoin('/course/', course.permalink)}>Lire le tutoriel</a>
-        </div>
-      </div>
-    );
-  }
-
   render() {
     return (
       <div className="posts">
-        {this.props.courses.map(this.courseRenderer)}
+        {this.props.courses.map(courseRenderer)}
       </div>
     );
   }
